modularized gui/ & modules/, moved Terminal frame to dedicated class, running from app.py
