+++ src/etc/bashScripts/johnInstall.sh
