+++ src/etc/bashScripts/lab10tmkkcc.sh
