+++ src/etc/bashScripts/newUser.sh
