--- src/service/test_service.py
@@ -1,2 +1,0 @@
-
-print("hello from the test service")
